refactor(types): consolidate IWorkerPool interface into types.ts
- Move IWorkerPool interface from workerPoolInterface.ts to types.ts
- Delete workerPoolInterface.ts
- Update imports in workerPool.ts, index.ts, and sed.ts
- Consolidate all type definitions in single location

All 149 tests pass.

implement #41
